move get_hint_if_single_char_arg to methods/utils.rs
